python_tdd/tests/test_ecs/my_ecs_construct/requirements.txt
aws-cdk-lib
constructs
jadecobra
sniffer
macfsevents